CORS-debug-mesh/requirements.txt
Flask
Pillow
requests
Flask-Cors
